sources/requirements.txt
Django==1.9.5
Flask==0.10.1
Flask-MySQL==1.3
Flask-RESTful==0.3.5
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.7.0
Twisted==13.2.0
Werkzeug==0.11.4
altgraph==0.10.2
aniso8601==1.1.0
bdist-mpkg==0.5.0
beautifulsoup4==4.5.0
bonjour-py==0.3
cassandra-driver==3.6.0
cql==1.4.0
dj-config-url==0.1.1
django-ajax==0.2.0
django-widget-tweaks==1.4.1
futures==3.0.5
itsdangerous==0.24
macholib==1.5.1
matplotlib==1.3.1
modulegraph==0.10.4
numpy==1.8.0rc1
peewee==2.8.1
psycopg2==2.6.1
py2app==0.7.3
pyOpenSSL==0.13.1
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
requests==2.9.1
scipy==0.13.0b1
six==1.10.0
thrift==0.9.3
whitenoise==3.2
wsgiref==0.1.2
xattr==0.6.4
zope.interface==4.1.1
